Reorganize tests
To pass tests on package built from source with pybuild:
- In tests/, replace tests that require access to original data files with tests on readme examples.
- Move all tests based on original data files to dev_tests.
